pyproject.toml
[project]
name = "chimera-bench"
version = "0.1.0"
dependencies = ["pyyaml"]

[tool.pytest.ini_options]
addopts = "-q"
